+++ packages/demoreact/src/app/SpaceUITest/SpaceBanner.tsx
@@ -6,7 +6,7 @@ import Dropdown from '../components/Dropdown';
 export const SpaceBanner = () => {
   const { SpaceBannerComponent } = useSpaceComponents();
   const [spaceId, setSpaceId] = useState<string>('');
-  const [orientation, setOrientation] = useState('');
+  const [orientation, setOrientation] = useState('maximized');
   const updateSpaceId = (e: React.SyntheticEvent<HTMLElement>) => {
     setSpaceId((e.target as HTMLInputElement).value);
   };
