+++ packages/uiweb/src/lib/context/spacesContext.ts
@@ -1,15 +1,6 @@
 import React, { createContext } from 'react';
 import { SpaceDTO } from '@pushprotocol/restapi';
 
-// export interface ISpaceBannerData {
-//   spaceId: string;
-//   apiData: {
-//     spaceName: string;
-//     members : any;
-//     pendingMembers : [];
-//     scheduleAt: string;
-//   };
-// }
 
 export interface ISpaceDataContextValues {
   trendingListData: any;
